==> Rcode/00.0_install_libraries.R <==
# Install required libraries

install.packages('DBI')
install.packages('dplyr')
install.packages('ggpubr')
install.packages('gridExtra')
install.packages('htm2txt')
install.packages('httr')
install.packages('jsonlite')
install.packages('knitr')
install.packages('lubridate')
install.packages('mgsub')
install.packages('RCurl')
install.packages('readxl')
install.packages('rmarkdown')
install.packages('RSQLite')
install.packages('rstudioapi')
install.packages('sf')
install.packages('stringr')
install.packages('tidybayes')
install.packages('tidyverse')
