==> Rcode/03.0_generate_plots.Rmd <==
---
title: "03.0_generate_plots.Rmd"
author: "Thomas Richards"
date: "26/08/2024"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
knitr::opts_chunk$set(dev = 'jpg')
knitr::opts_chunk$set(dpi=300)
```

# Generate Plots

This workbook generates plots from inhaler prescribing data and measures previously calculated.

## Libraries required:
stringr
tidyverse
ggpubr
gridExtra
tidybayes

## Data files required:

BNF_codes_measures.csv
OP_prescriptions.csv
IMD_all.csv
categories_prac.csv
categories_lsoa.csv
OP_measures_lsoa.csv

## Output files:

figure_1.jpg
figure_2.jpg
figure_3.jpg
figure_4.jpg


## 1. Load libraries and import data

Load libraries, import required data files, and set the working directory to the folder containing the workbook:


```{r}
library(stringr)
library(tidyverse)
library(ggpubr)
library(gridExtra)
library(tidybayes)

# get directory of current script and set it as working directory
filedir <- dirname(rstudioapi::getSourceEditorContext()$path)
setwd(filedir)
getwd()

# Output plot directory
output_dir_data <- './data/'
output_dir_plots <- './plots/'

# create data folder if not yet existent
if (!dir.exists("./plots")) {
  dir.create("./plots")
}

# load BNF code data
BNF_codes_measures <- read.csv("./data/BNF_codes_measures.csv")

# load the prescription data
OP_prescriptions <- read.csv("./data/OP_prescriptions.csv") %>%
  filter(setting == 4,
         bnf_code %in% BNF_codes_measures$bnf_code) %>%
  mutate(ccg = gsub('.0', 'E', ccg, fixed = TRUE),
         ccg = ifelse(ccg == 112, '16C', ccg),
         date = as.POSIXct(date, format="%Y-%m-%d"),
         year=format(date, "%Y"),
         month=format(date, "%m"),
         year_month=format(date, "%Y-%m"))

# load deprivation data
IMD_all <- read.csv("./data/IMD_all.csv")

# load counts of inhaler categories at practice level
categories_prac <- read.csv(paste(output_dir_data, 'categories_prac.csv', sep='')) %>%
  mutate(date = as.POSIXct(date, format="%Y-%m-%d"),
         year=format(date, "%Y"),
         month=format(date, "%m"),
         year_month=format(date, "%Y-%m"))

# load counts of inhaler categories at lsoa level
categories_lsoa <- read.csv(paste(output_dir_data, 'categories_lsoa.csv', sep='')) %>%
  mutate(date = as.POSIXct(date, format="%Y-%m-%d"),
         year=format(date, "%Y"),
         month=format(date, "%m"),
         year_month=format(date, "%Y-%m"))

# load prescribing measures at lsoa level
OP_measures_lsoa <- read.csv(paste(output_dir_data,"OP_measures_lsoa.csv", sep='')) %>%
  mutate(date = as.POSIXct(date, format="%Y-%m-%d"),
         year=format(date, "%Y"),
         month=format(date, "%m"),
         year_month=format(date, "%Y-%m"))

# start time to study
time_start <- "2018-12-01"

```

## Total number of prescriptions for inhaler subsets over time (Fig1B)
```{r fig.height=1.5, fig.width=3}
# plot total UK prescriptions by year
for_lines_subsets <- categories_prac %>%
  filter(date > time_start) %>%
  mutate(low_dose_ics_products = ics_not_laba - high_dose_ics_products,
         dpi = mdi_dpi - mdi) %>%
  rename(ics_low_dose = low_dose_ics_products,
         ics_high_dose = high_dose_ics_products,
         saba = saba_inhaler_products,
         laba = laba_products) %>%
  pivot_longer(cols = c(
    all_saba_ics_products,
    mdi,
    dpi,
    ics_low_dose,
    ics_high_dose,
    saba,
    laba)) %>%
  group_by(date, name) %>%
  summarise(value = sum(value)) %>%
  ungroup() %>%
  mutate(chart = if_else(name %in% c('all_saba_ics_products', 'mdi', 'dpi'), 'Totals', 'Subsets'),
         chart_label = TRUE,
         name = ifelse(name == 'all_saba_ics_products', 'all_saba_ics', name))


final_date <- tibble(date = c("2022-01-01", "2022-01-01"),
                     name = c('',''),
                     value = c(0,0),
                     chart = c('Totals', 'Subsets'),
                     chart_label = c(TRUE, TRUE))

# add extra room on x axis for labels
for_lines_subsets <- rbind(for_lines_subsets, 
                           final_date)

# replace underscores with spaces
for_lines_subsets <- for_lines_subsets %>%
  mutate(name = gsub("_", " ", name))

# generate plot
plt <- ggplot(for_lines_subsets, 
              aes(x= date,
                  y=value,
                  group=name))

lines_national <- plt +
  geom_line() +
  facet_wrap(~chart, scales = 'free_y', ncol = 2) +
  labs(x = "Date", 
       y = "Prescriptions written",
       color = "Subset") +
  theme_minimal() +
  labs(title = "B", hjust = 'left') +
  theme(plot.title = element_text(size=25)) +
  scale_y_continuous(labels=scales::unit_format(suffix="m",
                                                scale=0.000001, sep="", big.mark   = ",")) +
  geom_vline(xintercept = as.POSIXct("2020-03-16"), linetype=4) +
  geom_vline(xintercept = as.POSIXct("2020-11-05"), linetype=4) +
  theme(legend.position = "none") +
  geom_text(data = . %>% filter(chart_label, date == "2020-12-01"),
            aes(x=as.POSIXct("2021-01-01"),
                y = value,
                label = paste0(name)),
            size = 3,
            hjust = 'left') 

# plot and save
plot(lines_national)
# ggsave(paste(output_dir_plots, "figure1B.jpg", sep=""))


```


## Number of prescriptions for ICS in March by year (Fig1A)
```{r fig.height=3, fig.width=3}
# percentage change in number of prescriptions march 19 - 20
for_bars_march <- categories_prac %>% 
  filter(setting == 4,
         !year == '2015',
         month == "03") %>%
  select(year, ics_not_laba, laba_products, saba_inhaler_products) %>% # choose this to display each category
  pivot_longer(c(ics_not_laba, laba_products, saba_inhaler_products)) %>%
  group_by(year, name) %>%
  summarise(Number_of_prescriptions = sum(value))

# replace underscores with spaces
for_bars_march <- for_bars_march %>%
  mutate(name = gsub("_", " ", name))

# generate plot
plt <- ggplot(data = for_bars_march,
              mapping = aes(
                x=year,
                y = Number_of_prescriptions,
                fill = name))

bars_march_national <- plt  + geom_bar(
  stat="identity",
  width=0.5) +
  theme_minimal() +
  labs(title = "A", hjust = 'left') +
  theme(plot.title = element_text(size=25)) +
  labs(x = "Year", y = "Number of Items") +
  scale_y_continuous(labels=scales::unit_format(suffix="m",
                                                scale=0.000001, sep="", big.mark   = ","))+
  theme(axis.text = element_text(size=12),
        title = element_text(size=12)) +
  guides(fill=guide_legend(title="Inhaler subset")) +
  scale_fill_grey()

# plot and save
plot(bars_march_national)
# ggsave(paste(output_dir_plots, "figure1A.jpg", sep=""),
#        plot=bars_march_national, dpi = 300)

```

## Overall change by category (Fig1C)
```{r }
# plot % increase feb-march by subset
for_bars_perc <- for_lines_subsets %>%
  rename(Measure = name) %>%
  group_by(Measure) %>%
  summarise(percent_change = ((value[date == "2020-03-01"] - value[date == "2020-02-01"])/
                                (value[date == "2020-02-01"])*100)) %>%
  mutate(name = ifelse(Measure == 'all_saba_ics_products', 'all_saba_ics', Measure),
         chart = if_else(Measure %in% c('all_saba_ics', 'mdi', 'dpi'), 'Totals', 'Subsets')) %>%
  arrange(chart)

# replace underscores with spaces
for_bars_perc <- for_bars_perc %>%
  mutate(name = gsub("_", " ", name))

# generate plot
plt <- ggplot(data = for_bars_perc,
              mapping = aes(x=Measure,
                            y=percent_change))

bars_perc_national <- plt + geom_bar(stat="identity", width=0.5) +
  facet_wrap(~chart, 
             scales = 'free', 
             ncol = 2) +
  theme_minimal() +
  labs(x = "Subset", y = "% change Feb to March 2020") +
  labs(title = "C", hjust = 'left') +
  theme(plot.title = element_text(size=25)) +
  scale_y_continuous(labels=scales::unit_format(suffix="%")) +
  theme(axis.text = element_text(size=10)) +
  theme(axis.text.x=element_text(angle=45,hjust=0.95,vjust=0.2)) + 
  theme(legend.position = "none") +
  theme(axis.text.x = element_text(vjust=1))

# plot and save
plot(bars_perc_national)
# ggsave(paste(output_dir_plots, "figure1C.jpg", sep=""), 
#        plot=bars_perc_national, dpi = 300)

```


## Figure 1
```{r fig.height=6, fig.width=3}
figure_1 <- grid.arrange(bars_march_national,
                  lines_national,
                  bars_perc_national)

ggsave(paste(output_dir_plots, "figure_1.jpg", sep=""), 
       plot=figure_1,
       dpi = 300,
       width = 8,
       height = 10,
       units = 'in')

```






## Calcualte % increase at LSOA level by decile rank
```{r}
# calculate %change in ICS prescriptions per LSOA feb-mar 2019
perc_dif_lsoa_long <- categories_lsoa %>%
  filter(date == "2020-02-01"|
           date == "2020-03-01") %>% 
rename(ics_saba_laba =  all_saba_ics_products,
       ics = ics_not_laba,
       laba = laba_products,
       saba = saba_inhaler_products) %>%
  select(date, lsoa_id, ics_saba_laba, ics, laba, saba, patients_resp_lsoa) %>%
  pivot_longer(cols = c(ics_saba_laba, ics, laba, saba)) %>%
  mutate(value = value/patients_resp_lsoa) %>%
  pivot_wider(id_cols = c('lsoa_id', 'name'),
              names_from = c(date),
              names_prefix = "items_",
              values_from = value) %>%
  rename(items_feb = "items_2020-02-01",
         items_march = "items_2020-03-01") %>%
  group_by(lsoa_id, name) %>%
  summarise(
    perc_dif = ((items_march-items_feb)/items_feb)*100)

perc_dif_lsoa <- perc_dif_lsoa_long %>%
  pivot_wider(names_prefix = "perc_dif_",
              values_from = 'perc_dif')

```

## Correlations of % increase in prescribing with deprivation
Note: Higher IMD_P_rank = Higher levels of deprivation
      Higher IMD_decile_rank = LOWER levels of deprivation

```{r}
perc_dif_lsoa_IMD_long <- perc_dif_lsoa_long %>%
  right_join(IMD_all,
            by = "lsoa_id") %>%
  drop_na()

perc_samp <- perc_dif_lsoa_IMD_long %>%
  filter(name == 'ics_saba_laba')

# plot(perc_samp$IMD_P_rank, perc_samp$perc_dif)

perc_dif_lsoa_IMD_long %>%
  group_by(name) %>%
  summarise(r_value_IMD_score = cor(IMD_score, perc_dif,
                                    method = "pearson", use = "complete.obs"),
            r_value_IMD_rank = cor(IMD_rank, perc_dif,
                                    method = "pearson", use = "complete.obs"),
            r_value_IMD_P_rank = cor(IMD_P_rank, perc_dif,
                                    method = "pearson", use = "complete.obs"),
            r_value_IMD_decile_rank =  cor(IMD_decile_rank, perc_dif,
                                    method = "pearson", use = "complete.obs")) %>%
  ungroup()

```
Note: Higher IMD_P_rank = Higher levels of deprivation
      Higher IMD_decile_rank = LOWER levels of deprivation
      
      
## Percentage difference by LSOA decile (Fig2B)

```{r fig.height=6, fig.width=1.5}

perc_dif_decile <- perc_dif_lsoa_IMD_long %>%
  group_by(name, IMD_decile_rank) %>%
  summarise(perc_dif_mean = mean(perc_dif),
            perc_dif_median = median(perc_dif),
            perc_dif_sd = sd(perc_dif)) %>%
  ungroup() %>%
  rbind(perc_dif_lsoa_IMD_long %>%
          group_by(name) %>%
          summarise(perc_dif_mean = mean(perc_dif),
                    perc_dif_median = median(perc_dif),
                    perc_dif_sd = sd(perc_dif)) %>%
          ungroup() %>%
          mutate(IMD_decile_rank = 'All')) %>%
  mutate(IMD_decile_rank = factor(IMD_decile_rank,
                                    levels = c(as.character(1:10), 'All')))

# arrange
perc_dif_decile[order(perc_dif_decile$IMD_decile_rank),]


# for bars
perc_dif_decile_bars <- perc_dif_decile %>%
  filter(IMD_decile_rank %in% c('1','10','All')) %>%
  mutate(IMD_decile_rank = ifelse(IMD_decile_rank == '1', 
                                  'Highest dep.', IMD_decile_rank),
         IMD_decile_rank = ifelse(IMD_decile_rank == '10', 
                                  'Lowest dep.', IMD_decile_rank),
         IMD_decile_rank = ifelse(IMD_decile_rank == '11', 
                                  'National', IMD_decile_rank)) %>%
  mutate(IMD_decile_rank =
           factor(IMD_decile_rank, levels = c('Lowest dep.',
                                              'National',
                                              'Highest dep.')))

# replace underscores with spaces
perc_dif_decile_bars <- perc_dif_decile_bars %>%
  mutate(name = gsub("_", " ", name))

# generate plot
bars_IMD_med <- ggplot(perc_dif_decile_bars,
                       aes(x= IMD_decile_rank,
                           y=perc_dif_median)) +
  geom_bar(stat="identity",
           position = 'dodge',
           width=0.5) +
  facet_wrap(~name,
             # rows = vars(measure),
             scales="free_y",
             ncol = 1) +
  theme_minimal() +
  labs(x = "IMD decile rank", 
       y = "Median % increase in March 2020") +
  labs(title = "B", hjust = 'left') +
  theme(plot.title = element_text(size=25)) +
  theme(axis.text = element_text(size=10),
        title = element_text(size=10)) +
  theme(axis.text.x = element_text(angle = 35,
                                   hjust = .75))

# plot and save
plot(bars_IMD_med)
# ggsave(paste(output_dir_plots, "figure_2B.jpg", sep=""))

```


## Make categories_lsoa_long
```{r}
categories_lsoa_long <- categories_lsoa %>%
  rename(
    ics_laba_not_saba = all_ics_products,
    ics_saba_laba =  all_saba_ics_products,
    ics = ics_not_laba,
    laba = laba_products,
    saba = saba_inhaler_products) %>%
  select(date, lsoa_id, ics_saba_laba, ics, laba, saba,
         patients_resp_lsoa) %>%
  pivot_longer(cols = c(ics_saba_laba, ics, laba, saba))

```



## Categories ribbons IMD raw (Fig2A)
```{r fig.height=3, fig.width=1.5}
date_stretch = as.POSIXct("2021-05-15", format="%Y-%m-%d")

# plot total UK prescriptions by year
categories_rib_plot <- categories_lsoa_long %>%
  filter(date > '2018-12-01')%>%
  left_join(IMD_all %>% select(lsoa_id, IMD_decile_rank), by = 'lsoa_id') %>%
  drop_na() %>%
  rbind(tibble(date = rep(date_stretch, 12),
               lsoa_id = rep('', 12),
               patients_resp_lsoa = rep(0, 12),
               name = rep(unique(categories_lsoa_long$name), 3),
               value = rep(NA, 12),
               IMD_decile_rank = c(rep(1, 4), rep(10, 4), rep('All', 4))))

categories_rib_plot_IMD_med <- categories_rib_plot %>%
  group_by(date, name, IMD_decile_rank) %>%
  summarise(decile_median = median(value)) %>%
  rbind(tibble(date = rep(date_stretch, 12),
                     name = rep(unique(categories_rib_plot$name), 3),
                     decile_median = rep(0, 12),
                     decile_rank = rep(0, 12))) %>%
  mutate(lab_var = ifelse(IMD_decile_rank %in% c(1, 10), TRUE, FALSE),
         lab_text = ifelse(IMD_decile_rank == 11, TRUE, lab_var),
         lab_text = ifelse(IMD_decile_rank == 1, " Most deprived", ""),
         lab_text = ifelse(IMD_decile_rank == 10, " Least deprived", lab_text),
         lab_text = ifelse(IMD_decile_rank == 11, " National Median", lab_text))

# replace underscores with spaces
categories_rib_plot <- categories_rib_plot %>%
  mutate(name = gsub("_", " ", name))

# replace underscores with spaces
categories_rib_plot_IMD_med <- categories_rib_plot_IMD_med %>%
  mutate(name = gsub("_", " ", name))

# Draw ribbon plot
ribplot <- ggplot() +
  geom_lineribbon(data = categories_rib_plot %>% 
           group_by(date, name) %>%
           median_qi(value, .width = c(.5, .8, .95)),
         aes(x = date, y = value, ymin = .lower, ymax = .upper),
         linetype = "dashed",
         color="white",
         size = .8) +
  theme_minimal() +
  geom_line(data = categories_rib_plot_IMD_med %>%
              filter(IMD_decile_rank %in% c(1,10)),
            aes(x = date,
                y = decile_median,
                color = as.factor(IMD_decile_rank),
                group = IMD_decile_rank),
            size = .8) +
  facet_wrap(~name,
             scales="free_y",
             ncol = 1) +
  geom_vline(xintercept = as.POSIXct("2020-03-16"), linetype=4) +
  geom_vline(xintercept = as.POSIXct("2020-11-05"), linetype=4) +
  theme(legend.position = "none") +
  scale_fill_grey(start=.8, end=0.2) +
  geom_text(data = categories_rib_plot_IMD_med %>% 
              filter(lab_var, date == "2020-12-01"),
            aes(x=as.POSIXct("2020-12-01"),
                y = decile_median,
                label = paste0(lab_text)),
            size = 3,
            hjust = 'left') +
  labs(x = "Date",
       y = "Number of prescriptions (Median)") +
  labs(title = "A", hjust = 'left') +
  theme(plot.title = element_text(size=25))

# plot and save
plot(ribplot)
# ggsave(paste(output_dir_plots, "figure_2A.jpg", sep=""), ribplot)

```


## Figure 2
```{r fig.height=6, fig.width=3}
figure_2 <- grid.arrange(ribplot, bars_IMD_med,
                         layout_matrix = rbind(c(1,1,1,2,2),
                                               c(1,1,1,2,2),
                                               c(1,1,1,2,2)))



ggsave(paste(output_dir_plots, "figure_2.jpg", sep=""), 
       plot=figure_2,
       dpi = 300,
       width = 10,
       height = 8,
       units = 'in')

```










## Prescribing measure ribbons with IMD (Fig3A)
```{r fig.height=3, fig.width=1.5}
date_stretch = as.POSIXct("2021-05-15", format="%Y-%m-%d")
n_measure <- 5

measures_rib_plot <- OP_measures_lsoa %>%
  filter(date > "2018-12-01") %>%
  select(date, measure, org_id, calc_value) %>%
  rename(lsoa_id = org_id,
         value = calc_value) %>%
  right_join(IMD_all %>% select(lsoa_id, IMD_decile_rank),
             by = "lsoa_id") %>%
  drop_na() %>%
  rbind(tibble(date = rep(date_stretch, n_measure*3),
               lsoa_id = rep('', n_measure*3),
               value = rep(NA, n_measure*3),
               measure = c(rep(unique(OP_measures_lsoa$measure), 3)),
               IMD_decile_rank = c(rep(1, n_measure), rep(10, n_measure), rep('All', n_measure))))  

measures_rib_plot_IMD_med <- measures_rib_plot %>%
  group_by(date, measure, IMD_decile_rank) %>%
  summarise(decile_median = median(value)) %>%
  rbind(tibble(date = rep(date_stretch, n_measure*3),
                     measure = rep(unique(measures_rib_plot$measure), 3),
                     decile_median = rep(0, n_measure*3),
                     decile_rank = rep(0, n_measure*3))) %>%
  mutate(lab_var = ifelse(IMD_decile_rank %in% c(1, 10), TRUE, FALSE),
         lab_text = ifelse(IMD_decile_rank == 11, TRUE, lab_var),
         lab_text = ifelse(IMD_decile_rank == 1, " Most deprived", ""),
         lab_text = ifelse(IMD_decile_rank == 10, " Least deprived", lab_text),
         lab_text = ifelse(IMD_decile_rank == 11, " National Median", lab_text))


# replace underscores with spaces
measures_rib_plot <- measures_rib_plot %>%
  mutate(measure = gsub("_", " ", measure))

# replace underscores with spaces
measures_rib_plot_IMD_med <- measures_rib_plot_IMD_med %>%
  mutate(measure = gsub("_", " ", measure))


# generate ribbon plot
ribplot_meas <- ggplot() +
  geom_lineribbon(data = measures_rib_plot %>% 
           group_by(date, measure) %>%
           median_qi(value, .width = c(.5, .8, .95)),
         aes(x = date, y = value, ymin = .lower, ymax = .upper),
         linetype = "dashed",
         color="white",
         size = .8) +
  theme_minimal() +
  geom_line(data = measures_rib_plot_IMD_med %>% 
              filter(IMD_decile_rank %in% c(1,10)),
            aes(x = date,
                y = decile_median,
                color = as.factor(IMD_decile_rank),
                group = IMD_decile_rank),
                size = .8) +
  facet_wrap(~measure,
             scales="free_y",
             ncol = 1) +
  geom_hline(data = measures_rib_plot %>%
               filter(measure == 'ics pppm'),
             aes(yintercept = 0.417), 
             linetype=1,
             color = 'grey') +
  geom_vline(xintercept = as.POSIXct("2020-03-16"), linetype=4) +
  geom_vline(xintercept = as.POSIXct("2020-11-05"), linetype=4) +
  theme(legend.position = "none") +
  scale_fill_grey(start=.8, end=0.2) +
  geom_text(data = measures_rib_plot_IMD_med %>% 
              filter(lab_var, date == "2020-12-01"),
            aes(x=as.POSIXct("2020-12-01"),
                y = decile_median,
                label = paste0(lab_text)),
            size = 3,
            hjust = 'left') +
  labs(x = "Date",
       y = "Prescribing measure index (Median)") +
  labs(title = "A", hjust = 'left') +
  theme(plot.title = element_text(size=25))

# plot and save
plot(ribplot_meas)
# ggsave(paste(output_dir_plots, "figure_3A.jpg", sep=""), ribplot_meas)

```

## Calculate pre/lockdown/post and median for prescribing measures

```{r}
pre_dates <- seq(as.POSIXct("2019-04-01"), by = "month", length.out = 11)
post_dates <- seq(as.POSIXct("2020-04-01"), by = "month", length.out = 11)

measures_pre_post_lsoa <- measures_rib_plot %>%
  filter(date > "2019-03-01",
         date < "2021-01-01") %>%
  mutate(date_period = ifelse(date %in% pre_dates, "pre", ""),
         date_period = ifelse(date == as.Date("2020-03-01"), "lockdown", date_period),
         date_period = ifelse(date %in% post_dates, "post", date_period))

measures_pre_post_lsoa_median <- measures_pre_post_lsoa %>%
  rbind(measures_pre_post_lsoa %>%
          mutate(IMD_decile_rank = 'All')) %>%
  group_by(date_period, lsoa_id, measure, IMD_decile_rank) %>%
  summarise(period_avg = median(value)) %>%
  ungroup()

measures_pre_post_median <- measures_pre_post_lsoa_median %>%
  group_by(date_period, measure, IMD_decile_rank) %>%
  summarise(period_avg = median(period_avg)) %>%
  ungroup()

for_line_IMD_orig <- measures_pre_post_median %>%
  mutate(IMD_decile_rank = factor(IMD_decile_rank,
                                    levels = c(as.character(1:10), 'All'))) %>%
  filter(IMD_decile_rank %in% c('1','10','All')) %>%
  mutate(IMD_decile_rank = ifelse(IMD_decile_rank == '1', "  Most deprived", IMD_decile_rank),
         IMD_decile_rank = ifelse(IMD_decile_rank == '10', "  Least deprived", IMD_decile_rank),
         IMD_decile_rank = ifelse(IMD_decile_rank == '11', "  National Median", IMD_decile_rank))

```

## Line plot for pre/lockdown/post for IMD 1, 10, and overall (Fig3B)

```{r}
for_line_IMD <- for_line_IMD_orig %>%
  rbind(tibble(date_period = ' ',
               measure = 'ics_pppm',
               IMD_decile_rank = ' ',
               period_avg = NA)) %>%
  mutate(lab_var = ifelse(date_period == 'post', TRUE, '')) %>%
  mutate(date_period = factor(date_period, levels = c('pre','lockdown','post', ' ')))

# replace underscores with spaces
for_line_IMD <- for_line_IMD %>%
  mutate(measure = gsub("_", " ", measure))

# generate line plots
measlines_prepost <- ggplot(data = for_line_IMD,
                   aes(x=date_period, 
                       y=period_avg,
                       group = IMD_decile_rank,
                       linetype = IMD_decile_rank)) +
  geom_line() +
  geom_point() +
  theme_minimal() +
  facet_wrap(~measure, 
             scales="free_y",
             ncol = 1) +
  geom_hline(data = for_line_IMD %>% 
               filter(measure == 'ics pppm'),
             aes(yintercept = 0.417), linetype=1) +
  theme(legend.position = "none") +
  geom_text(data = for_line_IMD %>%
              filter(lab_var == TRUE),
            aes(x= "post",
                y = period_avg,
                label = paste0(IMD_decile_rank)),
            size = 3,
            hjust = 'left') +
  labs(x = "Time Period",
       y = "Prescribing index aggregated by time period") +
  labs(title = "B", hjust = 'left') +
  theme(plot.title = element_text(size=25))


# plot and save      
plot(measlines_prepost)
# ggsave(paste(output_dir_plots, "figure_3B.jpg", sep=""),
#        measlines_prepost)
```

## Figure 3
```{r fig.height=6, fig.width=3}
figure_3 <- grid.arrange(ribplot_meas, measlines_prepost,
                         layout_matrix = rbind(c(1,1,1,2,2),
                                               c(1,1,1,2,2),
                                               c(1,1,1,2,2),
                                               c(1,1,1,2,2)))

ggsave(paste(output_dir_plots, "figure_3.jpg", sep=""), 
       plot=figure_3,
       dpi = 300,
       width = 10,
       height = 10,
       units = 'in')


```




## Figure 4
## Environmental inhalers by generic/non-generic
```{r}
# sort data
env_inhalers <- BNF_codes_measures %>%
  filter(mdi_dpi == TRUE)

for_env_plot <- OP_prescriptions %>%
  filter(bnf_code %in% env_inhalers$bnf_code) %>%
  left_join(env_inhalers %>%
              select(bnf_code, bnf_name, is_generic, mdi, mdi_dpi),
            by = c('bnf_code', 'bnf_name')) %>%
  drop_na() %>%
  group_by(date, is_generic) %>%
  summarise(mdi = sum(items[mdi == "TRUE"]),
            mdi_dpi = sum(items[mdi_dpi == "TRUE"]),
            dpi = mdi_dpi - mdi) %>%
  ungroup() %>% 
  select(-mdi_dpi) %>%
  pivot_longer(cols = c(mdi, dpi),
               names_to = 'subset',
               values_to = 'n_items')

# summarise 
for_env_plot1 <- for_env_plot %>%
  group_by(date, subset) %>%
  summarise(n_items = sum(n_items)) %>%
  ungroup()

# generate plot 1
env_plot1 <- ggplot(data = for_env_plot1,
                    aes(x = date,
                        y = n_items,
                        group = subset,
                        color = subset,
                        fill = subset)) +
  geom_line(size=1) +
  theme_minimal() +
  labs(x = "Date",
       y = "Number of Items") +
  scale_y_continuous(labels=scales::unit_format(suffix="m",
                                                scale=0.000001, sep="", big.mark   = ",")) +
    scale_colour_manual(values = 
                          c(dpi = "#00BFC4",
                            mdi = "#F8766D"))+
  theme(axis.text = element_text(size=12),
        title = element_text(size=12)) +
  guides(color=guide_legend(title="")) +
  labs(title = "A", hjust = 'left') +
  theme(plot.title = element_text(size=25))



env_plot1

# Change "is_generic" options
for_env_plot <- for_env_plot %>%
  mutate(is_generic = ifelse(is_generic == TRUE, "Generic", is_generic),
         is_generic = ifelse(is_generic == FALSE, "Non-generic", is_generic))

# generate plot 2
env_plot2 <- ggplot(data = for_env_plot,
                    aes(x = date,
                        y = n_items,
                        group = is_generic,
                        color = is_generic,
                        fill = is_generic
                        )) +
  facet_grid(~subset) +
  geom_line(size=1) +
  theme_minimal() +
  labs(x = "Date",
       y = "Number of Items") +
  scale_y_continuous(labels=scales::unit_format(suffix="m",
                                                scale=0.000001, sep="", big.mark   = ",")) +
  scale_color_manual(values = c(Generic = "#00BFC4",
                                 "Non-generic" = "#F8766D"))+
  theme(axis.text = element_text(size=12),
        title = element_text(size=12),
        strip.text.x = element_text(size = 14)) +
  guides(color=guide_legend(title="")) +
  labs(title = "B", hjust = 'left') +
  theme(plot.title = element_text(size=25))


env_plot2

# combine plots and save
figure_env <- grid.arrange(env_plot1, env_plot2)
ggsave(paste(output_dir_plots, "figure_4.jpg", sep=""), 
       plot=figure_env,
       dpi = 300,
       width = 6,
       height = 6,
       units = 'in')


```
